refactor(core): Move Git-related modules to dedicated core/git directory
