+++ src/cli/actions/remoteAction.ts
@@ -1,17 +1,15 @@
 import * as fs from 'node:fs/promises';
 import os from 'node:os';
 import path from 'node:path';
-import GitUrlParse, { type GitUrl } from 'git-url-parse';
 import pc from 'picocolors';
-import { execGitShallowClone, isGitInstalled } from '../../core/file/gitCommand.js';
+import { execGitShallowClone, isGitInstalled } from '../../core/git/gitCommand.js';
+import { parseRemoteValue } from '../../core/git/gitRemoteParse.js';
 import { RepomixError } from '../../shared/errorHandle.js';
 import { logger } from '../../shared/logger.js';
 import { Spinner } from '../cliSpinner.js';
 import type { CliOptions } from '../types.js';
 import { type DefaultActionRunnerResult, runDefaultAction } from './defaultAction.js';
-interface IGitUrl extends GitUrl {
-  commit: string | undefined;
-}
+
 export const runRemoteAction = async (
   repoUrl: string,
   cliOptions: CliOptions,
@@ -53,69 +51,6 @@ export const runRemoteAction = async (
   }
 
   return result;
-};
-
-// Check the short form of the GitHub URL. e.g. yamadashy/repomix
-const VALID_NAME_PATTERN = '[a-zA-Z0-9](?:[a-zA-Z0-9._-]*[a-zA-Z0-9])?';
-const validShorthandRegex = new RegExp(`^${VALID_NAME_PATTERN}/${VALID_NAME_PATTERN}$`);
-export const isValidShorthand = (remoteValue: string): boolean => {
-  return validShorthandRegex.test(remoteValue);
-};
-
-export const parseRemoteValue = (remoteValue: string): { repoUrl: string; remoteBranch: string | undefined } => {
-  if (isValidShorthand(remoteValue)) {
-    logger.trace(`Formatting GitHub shorthand: ${remoteValue}`);
-    return {
-      repoUrl: `https://github.com/${remoteValue}.git`,
-      remoteBranch: undefined,
-    };
-  }
-
-  try {
-    const parsedFields = GitUrlParse(remoteValue) as IGitUrl;
-
-    // This will make parsedFields.toString() automatically append '.git' to the returned url
-    parsedFields.git_suffix = true;
-
-    const ownerSlashRepo =
-      parsedFields.full_name.split('/').length > 1 ? parsedFields.full_name.split('/').slice(-2).join('/') : '';
-
-    if (ownerSlashRepo !== '' && !isValidShorthand(ownerSlashRepo)) {
-      throw new RepomixError('Invalid owner/repo in repo URL');
-    }
-
-    const repoUrl = parsedFields.toString(parsedFields.protocol);
-
-    if (parsedFields.ref) {
-      return {
-        repoUrl: repoUrl,
-        remoteBranch: parsedFields.filepath ? `${parsedFields.ref}/${parsedFields.filepath}` : parsedFields.ref,
-      };
-    }
-
-    if (parsedFields.commit) {
-      return {
-        repoUrl: repoUrl,
-        remoteBranch: parsedFields.commit,
-      };
-    }
-
-    return {
-      repoUrl: repoUrl,
-      remoteBranch: undefined,
-    };
-  } catch (error) {
-    throw new RepomixError('Invalid remote repository URL or repository shorthand (owner/repo)');
-  }
-};
-
-export const isValidRemoteValue = (remoteValue: string): boolean => {
-  try {
-    parseRemoteValue(remoteValue);
-    return true;
-  } catch (error) {
-    return false;
-  }
 };
 
 export const createTempDirectory = async (): Promise<string> => {
