+++ src/core/git/gitCommand.ts
