+++ src/core/git/gitRemoteParse.ts
@@ -1,0 +1,70 @@
+import gitUrlParse, { type GitUrl } from 'git-url-parse';
+import { RepomixError } from '../../shared/errorHandle.js';
+import { logger } from '../../shared/logger.js';
+
+interface IGitUrl extends GitUrl {
+  commit: string | undefined;
+}
+
+// Check the short form of the GitHub URL. e.g. yamadashy/repomix
+const VALID_NAME_PATTERN = '[a-zA-Z0-9](?:[a-zA-Z0-9._-]*[a-zA-Z0-9])?';
+const validShorthandRegex = new RegExp(`^${VALID_NAME_PATTERN}/${VALID_NAME_PATTERN}$`);
+export const isValidShorthand = (remoteValue: string): boolean => {
+  return validShorthandRegex.test(remoteValue);
+};
+
+export const parseRemoteValue = (remoteValue: string): { repoUrl: string; remoteBranch: string | undefined } => {
+  if (isValidShorthand(remoteValue)) {
+    logger.trace(`Formatting GitHub shorthand: ${remoteValue}`);
+    return {
+      repoUrl: `https://github.com/${remoteValue}.git`,
+      remoteBranch: undefined,
+    };
+  }
+
+  try {
+    const parsedFields = gitUrlParse(remoteValue) as IGitUrl;
+
+    // This will make parsedFields.toString() automatically append '.git' to the returned url
+    parsedFields.git_suffix = true;
+
+    const ownerSlashRepo =
+      parsedFields.full_name.split('/').length > 1 ? parsedFields.full_name.split('/').slice(-2).join('/') : '';
+
+    if (ownerSlashRepo !== '' && !isValidShorthand(ownerSlashRepo)) {
+      throw new RepomixError('Invalid owner/repo in repo URL');
+    }
+
+    const repoUrl = parsedFields.toString(parsedFields.protocol);
+
+    if (parsedFields.ref) {
+      return {
+        repoUrl: repoUrl,
+        remoteBranch: parsedFields.filepath ? `${parsedFields.ref}/${parsedFields.filepath}` : parsedFields.ref,
+      };
+    }
+
+    if (parsedFields.commit) {
+      return {
+        repoUrl: repoUrl,
+        remoteBranch: parsedFields.commit,
+      };
+    }
+
+    return {
+      repoUrl: repoUrl,
+      remoteBranch: undefined,
+    };
+  } catch (error) {
+    throw new RepomixError('Invalid remote repository URL or repository shorthand (owner/repo)');
+  }
+};
+
+export const isValidRemoteValue = (remoteValue: string): boolean => {
+  try {
+    parseRemoteValue(remoteValue);
+    return true;
+  } catch (error) {
+    return false;
+  }
+};
