+++ src/core/packager.ts
@@ -5,7 +5,7 @@ import { sortPaths } from './file/filePathSort.js';
 import { processFiles } from './file/fileProcess.js';
 import { searchFiles } from './file/fileSearch.js';
 import type { RawFile } from './file/fileTypes.js';
-import { GitDiffResult, getGitDiffs } from './file/gitDiff.js';
+import { GitDiffResult, getGitDiffs } from './git/gitDiff.js';
 import { calculateMetrics } from './metrics/calculateMetrics.js';
 import { generateOutput } from './output/outputGenerate.js';
 import { copyToClipboardIfEnabled } from './packager/copyToClipboardIfEnabled.js';
